+++ life_expectancy/cleaning.py
@@ -1,24 +1,6 @@
 """Module for cleaning life expectancy data."""
 
-import argparse
-from pathlib import Path
-from typing import Optional
 import pandas as pd
-
-BASE_DIR = Path().cwd() / 'life_expectancy' / 'data'
-
-def load_data(path: Optional[Path] = None) -> pd.DataFrame:
-    """ Read life expectancy data into a DataFrame.
-    
-    Args:
-        path (Optional[Path]): The path to the directory with the csv file.
-
-    Returns:
-        (pd.DataFrame): The data.
-    """
-    if path is None:
-        path = BASE_DIR / 'eu_life_expectancy_raw.tsv'
-    return pd.read_csv(path, sep='\t')
 
 def split_first_column(df: pd.DataFrame) -> pd.DataFrame:
     """ Split four variables in the first column into their own columns.
@@ -130,37 +112,3 @@ def clean_data(df: pd.DataFrame, region: str = 'PT') -> pd.DataFrame:
     df = correct_data_types(df)
     df = df.dropna()
     return filter_region_data(df, region)
-
-def save_data(df: pd.DataFrame, region: str, path: Optional[Path] = None) -> None:
-    """ Save data for the chosen region.
-    Args:
-        df (pd.DataFrame): The data.
-        region (str): The region to select.
-        path (Optional[Path]): The path to the export directory.
-    """
-    file_name = f'{region.lower()}_life_expectancy.csv'
-    if path is None:
-        path = BASE_DIR / file_name
-    else:
-        path = path / file_name
-    df.to_csv(path, index = False)    
-
-def main(region: str)-> pd.DataFrame:
-    """Main function.
-    
-    Args:
-        region (str): The region to select.
-    """
-    eu_life_expectancy_data = load_data()
-    region_life_expectancy_data = clean_data(eu_life_expectancy_data, region = region)
-    save_data(region_life_expectancy_data, region)
-    return region_life_expectancy_data
-
-if __name__ == "__main__":  # pragma: no cover
-    parser = argparse.ArgumentParser()
-    parser.add_argument("-r", "--region",
-                         default = 'PT', 
-                         help = 'Input region argument.',
-                         type = str)
-    args = parser.parse_args()
-    main(region = args.region)
